requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
langchain-groq==0.1.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.4.0
tqdm==4.66.1
python-json-logger==2.0.7
PyYAML==6.0.1
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
tenacity==8.2.3
scipy==1.10.1
joblib==1.3.2